fix: Token check fix.

+++ controllers/api/graph.js
@@ -6,39 +6,14 @@ const db = require("../../database/");
 const server = require("../../socket");
 const apiTools = require("../../lib/apiTools");
 const ApiReport = require("../../lib/ApiReport");
-const { getRelatedNodes } = require("../../database/api/graph");
 
-/**
- * 
- * @param {any} req 
- * @param {any} res 
- *//*
-module.exports.createMark = (req, res) => apiTools.parameterizedHandler(["type"], req, res, (obj, req, res) => {
-    db.graph.createNode(["Mark"], obj)
-        .then()
-        .catch();
-    console.log([1, 2, 3] instanceof Array);
-});*/
-/*
-{
-    type: "type_name",
-    properties: 
-    [
-        {
-            propertyName: "prop1"
-            require: true | false
-            default: "str" | number
-        },
-        {
-            propertyName: "prop2"
-            require: true | false
-            default: "str" | number
-        }
-    ]
-}
-*/
+module.exports.createMark = apiTools.parameterizedHandler(["token", "data"], async function body(obj, req, res) {
+    if(!(await db.user.isTokenExists(obj.token)))
+    {
+        apiTools.sendReport(res, new ApiReport("error", 999, "Wrong token!"));
+        return;
+    }
 
-module.exports.createMark = apiTools.parameterizedHandler(["data"], async function body(obj, req, res) {
     let data;
 
     try {
@@ -76,26 +51,51 @@ module.exports.createMark = apiTools.parameterizedHandler(["data"], async functi
     apiTools.sendReport(res, new ApiReport("ok", 0, "Successful!"));
 }); 
 
-module.exports.createRelation = apiTools.parameterizedHandler(["from", "to", "name"], (obj, req, res) => {
+module.exports.createRelation = apiTools.parameterizedHandler(["token", "from", "to", "name"], async function body(obj, req, res) {
+    if(!(await db.user.isTokenExists(obj.token)))
+    {
+        apiTools.sendReport(res, new ApiReport("error", 999, "Wrong token!"));
+        return;
+    }
+
     db.graph.createRelation(obj.from, obj.to, obj.name);
+    apiTools.sendReport(res, new ApiReport("ok", 0, "Successful!"));
 });
 
-module.exports.deleteNode = apiTools.parameterizedHandler(["id"], (obj, req, res) => {
+module.exports.deleteNode = apiTools.parameterizedHandler(["token", "id"], async function body(obj, req, res) {
+    if(!(await db.user.isTokenExists(obj.token)))
+    {
+        apiTools.sendReport(res, new ApiReport("error", 999, "Wrong token!"));
+        return;
+    }
+
     db.graph.deleteNode(obj.id);
     apiTools.sendReport(res, new ApiReport("ok", 0, "Successful!"));
 });
 
-module.exports.deleteRelation = apiTools.parameterizedHandler(["from", "to", "name"], (obj, req, res) => {
+module.exports.deleteRelation = apiTools.parameterizedHandler(["token", "from", "to", "name"], async function body(obj, req, res) {
+    if(!(await db.user.isTokenExists(obj.token)))
+    {
+        apiTools.sendReport(res, new ApiReport("error", 999, "Wrong token!"));
+        return;
+    }
+
     db.graph.deleteRelation(obj.from, obj.to, obj.name);
     apiTools.sendReport(res, new ApiReport("ok", 0, "Successful!"));
 });
 
-module.exports.getMarksInfo = async function body(req, res) {
+module.exports.getMarksInfo = apiTools.parameterizedHandler(["token"], async function body(obj, req, res) {
+    if(!(await db.user.isTokenExists(obj.token)))
+    {
+        apiTools.sendReport(res, new ApiReport("error", 999, "Wrong token!"));
+        return;
+    }
+
     let response = [];
     let marks = await db.graph.getNodes("Mark");
     for (let i = 0; i < marks.records.length; i++)
     {
-        let relatedNodes = await getRelatedNodes(marks.records[i]["_fields"][0]["identity"]["low"], "property");
+        let relatedNodes = await db.graph.getRelatedNodes(marks.records[i]["_fields"][0]["identity"]["low"], "property");
 
         let mark = {};
         mark["type"] = marks.records[i]["_fields"][0]["properties"]["type"];
@@ -110,5 +110,5 @@ module.exports.getMarksInfo = async function body(req, res) {
     }
 
     apiTools.sendReport(res, new ApiReport("ok", 0, "Successful!", {response: response}));
-};
+});
